% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year}
\alias{check_year}
\title{Check year format}
\usage{
check_year(x)
}
\arguments{
\item{x}{year in Gregorian or Japanese calendar (eg "R2" and "\u5e73\u6210\u5143\u5e74\u5ea6")}
}
\value{
year in Gregorian calendar
}
\description{
Function to check and validate year
}
