% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_code_pref_muni}
\alias{df_code_pref_muni}
\title{Prefecture/Municipality Codes and Names}
\format{
## `df_code_pref_muni`
A data frame with 1,965 rows and 7 columns:
\describe{
  \item{code_dantai}{Dantai Code (5 digits)}
  \item{code_region}{Region Code}
  \item{code_pref}{Prefecture Code}
  \item{code_muni}{Municipality Code}
  \item{name_pref}{Prefecture Name}
  \item{name_muni}{Municipality Name}
  \item{name_muni}{Municipality Name in Hiragana}
}
}
\source{
<https://www.soumu.go.jp/denshijiti/code.html>
}
\usage{
df_code_pref_muni
}
\description{
Prefecture/Municipality Codes and Names
}
\keyword{datasets}
