% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_muni_mesh1}
\alias{df_muni_mesh1}
\title{Prefecture/Municipality/Mesh Codes at Mesh Level 1}
\format{
## `df_muni_mesh1`
A data frame with 2,922 rows and 4 columns:
\describe{
  \item{code_pref}{Prefecture Code}
  \item{code_muni}{Municipality Code}
  \item{code_mesh1}{Mesh Code (Level 1)}
}
}
\source{
<https://www.stat.go.jp/data/mesh/m_itiran.html>
}
\usage{
df_muni_mesh1
}
\description{
Prefecture/Municipality/Mesh Codes at Mesh Level 1
}
\keyword{datasets}
