% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_muni_mesh3}
\alias{df_muni_mesh3}
\title{Prefecture/Municipality/Mesh Codes at Mesh Level 3}
\format{
## `df_muni_mesh3`
A data frame with 2,922 rows and 4 columns:
\describe{
  \item{code_pref}{Prefecture Code}
  \item{code_muni}{Municipality Code}
  \item{name_muni}{Municipality Name}
  \item{code_mesh1}{Mesh Code (Level 1)}
  \item{code_mesh2}{Mesh Code (Level 2)}
  \item{code_mesh3}{Mesh Code (Level 3)}
}
}
\source{
<https://www.stat.go.jp/data/mesh/m_itiran.html>
}
\usage{
df_muni_mesh3
}
\description{
Prefecture/Municipality/Mesh Codes at Mesh Level 3
}
\keyword{datasets}
