% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_geojson_file}
\alias{find_geojson_file}
\title{Check year format}
\usage{
find_geojson_file(maptype, code_pref, code_muni, year, data_dir)
}
\arguments{
\item{x}{year in Gregorian or Japanese calendar (eg "R2" and "平成元年度")}
}
\value{
an epsg code
}
\description{
Function to check and validate year
}
