% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{get_tile_server}
\alias{get_tile_server}
\title{Provides URL of tile servers for base map}
\usage{
get_tile_server(map_type = "standard")
}
\arguments{
\item{map_type}{c("standard", "pale", "English", "photo")}
}
\value{
A list of characters containing server and reference information
}
\description{
Provides URL of tile servers for base map. "standard" and "light"
services have zoom levels 2 to 18.
For details, check:
https://maps.gsi.go.jp/development/ichiran.html
}
