% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo.R
\name{list_landnuminfo}
\alias{list_landnuminfo}
\title{message the list of available Land Numerical Information data.}
\usage{
list_landnuminfo()
}
\description{
message the list of available Land Numerical Information data.
}
