% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{read_census_odcity}
\alias{read_census_odcity}
\title{Download spatial data of census tracts of the Japanese Population Census}
\usage{
read_census_odcity(city_name, year = 2020)
}
\arguments{
\item{city_name}{an Ordnance-Disignated City in Japanese.}

\item{year}{Year of the data. Defaults to 2020.}
}
\value{
An `"sf" "data.frame"` object
}
\description{
Download spatial data of census tracts of the Japanese Population Census
of an Ordnance-Disignated City of all wards (ku) combined.
}
