% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{read_census_tract}
\alias{read_census_tract}
\title{Download spatial data of census tracts of the Japanese Population Census}
\usage{
read_census_tract(code_pref, code_muni, year = 2020, data_dir = NULL)
}
\arguments{
\item{code_pref}{The 2-digit code of a prefecture.}

\item{code_muni}{The 3-digit code of a municipality (city, town, or village).}

\item{year}{Year of the data. Defaults to 2020.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object
}
\description{
Download spatial data of census tracts of the Japanese Population Census
}
