% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo.R
\name{read_landnuminfo}
\alias{read_landnuminfo}
\title{Download spatial data of National Lang Numeric Information of Japan}
\usage{
read_landnuminfo(
  maptype,
  code_pref,
  code_muni = NULL,
  year = 2020,
  filetype = "geojson",
  geometry = NULL,
  data_dir = NULL,
  maptypeextra = ""
)
}
\arguments{
\item{maptype}{The map type code (e.g. "A01").}

\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{The 3-digit code of municipality (city, town, or village).}

\item{year}{Year of the data. Defaults to 2020.}

\item{filetype}{either "geojson" or "shp".}

\item{geometry}{"POINT", "LINESTRING", or "POLYGON"}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}

\item{maptypeextra}{File name suffix.}
}
\value{
An `"sf" "data.frame"` object
}
\description{
Low level function to download spatial data of National Lang Numeric Information
(or kokudo suuchi joho in Japanese) of Japan.
}
