% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo_landuse.R
\name{read_landnuminfo_forestarea_privateowned}
\alias{read_landnuminfo_forestarea_privateowned}
\title{Download spatial data of Forest Area (Private Owned) of Japan}
\usage{
read_landnuminfo_forestarea_privateowned(
  code_pref,
  code_muni = NULL,
  year = 2011,
  data_dir = NULL
)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Ignored.}

\item{year}{Year of the data. Defaults to 2011.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Forest Area (Private Owned). The returned value is an sf object with extra
attr "col" and "palette". The "col" is the factored column that indicate the land use classes
and "palette" provides the colour palette based on Japan Industrial Standard.
}
