% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo_water.R
\name{read_landnuminfo_lake}
\alias{read_landnuminfo_lake}
\title{Download spatial data of Lakes of Japan}
\usage{
read_landnuminfo_lake(
  code_pref = NULL,
  code_muni = NULL,
  year = 2005,
  data_dir = NULL
)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality.}

\item{year}{Year of the data. Defaults to 2005.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Lakes of Japan. The returned value is an sf object.
}
