% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo_mesh.R
\name{read_landnuminfo_mesh3}
\alias{read_landnuminfo_mesh3}
\title{Download spatial data of Mesh 3 of Japan}
\usage{
read_landnuminfo_mesh3(code_pref, code_muni, year = 2016, data_dir = NULL)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality.}

\item{year}{Year of the data. Defaults to 2016.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Mesh 3 of Japan. The returned value is an sf object.
}
