% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo_mesh.R
\name{read_landnuminfo_meshsub}
\alias{read_landnuminfo_meshsub}
\title{Download spatial data of Mesh Subdivision of Japan}
\usage{
read_landnuminfo_meshsub(code_pref, code_muni, year = 2006, data_dir = NULL)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality.}

\item{year}{Year of the data. Defaults to 2006. The years 2016,2014,2009,2006 are clean. Some mesh codes not available for other years.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Mesh Subdivision of Japan. The returned value is an sf object.
The function reads Land Use Subdivision Level Mesh data. See https://nlftp.mlit.go.jp/ksj/gml/datalist/KsjTmplt-L03-a.html.

Beware! The object size will be large! More than 620,000 rows! Use read_landnuminfo_mesh3() instead.

The available years are: 2021, 2016, 2014, 2009, 2006, 1997, 1991, 1987, 1976.

If the file of a mesh is not available at the site, then the mesh is simply ignored.

The default year is 2006 because the year covers isolated islands.
}
