% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo.R
\name{read_landnuminfo_preflandprice}
\alias{read_landnuminfo_preflandprice}
\title{Download spatial data of Prefecture-surveyed Land Prices of Japan}
\usage{
read_landnuminfo_preflandprice(
  code_pref,
  code_muni = NULL,
  year = 2023,
  data_dir = NULL
)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality.}

\item{year}{Year of the data. Defaults to 2022.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Prefecture-surveyed Land Prices of Japan. The returned value is an sf object.

The organisation defines this data as "PrefectureLandPriceResearch," but we call "Prefecture-surveyed Land Priced" instead.

Please note that some columns of real number (L02_033, L02_034) are coerced to integer for several years (e.g. 2018).
}
