% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo.R
\name{read_landnuminfo_urbanarea}
\alias{read_landnuminfo_urbanarea}
\title{Download spatial data of Urbanized Areas of Japan}
\usage{
read_landnuminfo_urbanarea(code_pref, code_muni, year = 2018, data_dir = NULL)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality. If specified, subtract the data by the column A48_003.}

\item{year}{Year of the data. Defaults to 2012.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" for tmap.
}
\description{
Function to download spatial data of Urbanized Areas of Japan. The returned value is an sf object. The license for year 2018 is open; or non-commercial for 2011 and 2006.
}
