% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo.R
\name{read_landnuminfo_welfare}
\alias{read_landnuminfo_welfare}
\title{Download spatial data of Welfare Facilities of Japan}
\usage{
read_landnuminfo_welfare(code_pref, code_muni, year = 2021, data_dir = NULL)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{The 3-digit code of municipality (city, town, or village).}

\item{year}{Year of the data. Defaults to 2012.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Welfare Facilities of Japan. The returned value is an sf object.
}
