% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhlw_ltci.R
\name{read_mhlw_ltci}
\alias{read_mhlw_ltci}
\title{Download spatial data of Long-Term Care Insurance services}
\usage{
read_mhlw_ltci(
  code_type = NULL,
  name_type = NULL,
  code_pref = NULL,
  code_muni = NULL,
  year = 2023,
  month = 6,
  data_dir = NULL,
  exclude_if_outside = FALSE
)
}
\arguments{
\item{code_type}{LTCI service code, e.g. 110 = Home-Visit Care.  See attribute sourceURL for details.}

\item{name_type}{LTCI service name in Japanese.}

\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality.}

\item{year}{Year of the data.}

\item{month}{Month of the data. Either 6 or 12.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Long-Term Care Insurance services,
supplied by Ministry of Health, Labour, and Welfare of Japan. The returned
value is an sf object.

The available years are: 2023, 2022, 2021 and 2020 (month = 12 only for 2020).
The default year is 2023.
}
