% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{year_2digit}
\alias{year_2digit}
\title{Converts 4-digit year to 2-digit}
\usage{
year_2digit(year)
}
\arguments{
\item{year}{year in 4-digit}
}
\value{
year in 2-digit as character
}
\description{
Converts 4-digit year to 2-digit
}
