% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_geojson_file}
\alias{find_geojson_file}
\title{Find the target geojson file}
\usage{
find_geojson_file(
  maptype,
  code_pref,
  code_muni,
  year,
  data_dir,
  maptypeextra = ""
)
}
\arguments{
\item{maptype}{landnuminfo maptype (eg "A01" and "P01").}

\item{code_pref}{The 2-digit code of a prefecture.}

\item{code_muni}{The 3-digit code of a municipality (city, town, or village).}

\item{year}{Year of the data. Defaults to 2012.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}

\item{maptypeextra}{File name suffix.}
}
\value{
the target filepath
}
\description{
Function to find the target geojson file in data_dir.
}
