% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landnuminfo.R
\name{read_landnuminfo_flood}
\alias{read_landnuminfo_flood}
\title{Download spatial data of Flood Inundation Risk of Japan}
\usage{
read_landnuminfo_flood(
  code_pref,
  code_muni = NULL,
  year = 2012,
  data_dir = NULL
)
}
\arguments{
\item{code_pref}{The 2-digit code of prefecture.}

\item{code_muni}{Optional. The 3-digit code of municipality.}

\item{year}{Year of the data. Defaults to 2012.}

\item{data_dir}{The directory to store downloaded zip and extracted files. If not specified, the data will be stored in a temp directory and will be deleted after you quit the session.}
}
\value{
An `"sf" "data.frame"` object with extra attr "col" and "palette" for tmap.
}
\description{
Function to download spatial data of Flood Inundation Risk of Japan. The returned value is an sf object.

Note that the data of year 2012 is available for all prefectures, but not of other years.
}
